@@ -18,15 +18,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
+    "@angular/animations": "^9.0.0",
     "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/common": "^9.0.0",
+    "@angular/compiler": "^9.0.0",
+    "@angular/core": "^9.0.0",
+    "@angular/forms": "^9.0.0",
+    "@angular/platform-browser": "^9.0.0",
+    "@angular/platform-browser-dynamic": "^9.0.0",
+    "@angular/router": "^9.0.0",
     "@microsoft/signalr": "^3.1.0",
     "@swimlane/ngx-charts": "^13.0.2",
     "@types/file-saver": "2.0.1",
@@ -41,15 +41,16 @@
     "ngx-md": "^8.1.6",
     "npm": "^6.13.7",
     "rxjs": "^6.5.4",
+    "tslib": "^1.10.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular/cli": "^8.3.23",
-    "@angular/compiler-cli": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular/cli": "^9.0.1",
+    "@angular/compiler-cli": "^9.0.0",
     "@types/jasmine": "^3.5.2",
-    "@types/node": "13.1.6",
-    "codelyzer": "^5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "core-js": "^3.6.4",
     "expose-loader": "^0.7.5",
     "imports-loader": "^0.8.0",
@@ -69,6 +70,6 @@
     "script-loader": "^0.7.2",
     "ts-node": "^8.6.2",
     "tslint": "^5.20.1",
-    "typescript": "~3.5.3"
+    "typescript": "~3.7.5"
   }
 }
