@@ -44,7 +44,7 @@ import { NgxMdModule } from 'ngx-md';
     ],
 })
 export class AuthModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AuthModule> {
     return {
       ngModule: AuthModule,
       providers: [AuthService]
